backend/python/requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Machine learning
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Database
pymongo==4.6.1
motor==3.3.2

# AI/ML models
openai==1.6.1
google-generativeai==0.3.2
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Async processing
asyncio-mqtt==0.13.0
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
click==8.1.7
httpx==0.26.0

# Data validation
marshmallow==3.20.1
cerberus==1.3.5

# Time series analysis
statsmodels==0.14.1
prophet==1.1.5

# Clustering and dimensionality reduction
umap-learn==0.5.5
hdbscan==0.8.33

# Performance monitoring
psutil==5.9.7
memory-profiler==0.61.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0 